JIB -> Dockerfile
The buildDockerMulti.sh script will build a docker image for both arm
and amd64 and push it to the registry. It will also create a multi
architecture manifest so that you don't need to manually specify which
arch you are want when pulling the image. Nothe that this requires you
to enable experimental docker features at the moment...

Also allow access to /actuator/health from localhost since this is used
for the health check in the image